added some more math to vector3, and worked on physics
did gravity and got a start on drag. need to ask for explanation and conformation on how the forces work.

added a namespace for vector3 math "Vmath" which has a function of magnitude, normalizing, and applying a power to the vector3.

got some more math that may be handy to add like dot product, or cross product.

I forgot to save the vector3.h so i will need to redo all math besides pow3